templates/tensorrt-6.0-cuda10.0/requirements.txt
algorithmia>=1.0.0,<2.0
six
numpy>=1.16.0,<2.0
pycuda==2018.1.1
Pillow>=5.2.0
wget>=3.2
tensorrt==6.0.1.5 # unchangable
onnx==1.5.0